pyproject.toml
[project]
name = "sharpqa-agent"
version = "0.1.0"
description = "SharpQA Sales Agent — local-first lead sourcing, website analysis, and cold outreach drafting"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "aiosqlite>=0.19",
    # HTTP & scraping
    "httpx>=0.27",
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    # LLM & embeddings
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "onnxruntime>=1.17,<1.24",
    "google-genai>=0.1.1",
    # API & dashboard
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "streamlit>=1.30",
    "sse-starlette>=2.0",
    # Scheduling
    "apscheduler>=3.10,<4",
    # Export
    "openpyxl>=3.1",
    # Utilities
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "respx>=0.21",
    "ruff>=0.3",
]

[project.scripts]
sharpqa = "sharpqa_agent.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["slow: marks tests that hit real network"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sharpqa_agent", "config"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
